# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Umrah Travel ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clients/{id}/notes:
    get:
      operationId: listClientNotes
      tags: [clients]
      summary: List notes for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientNote"
    post:
      operationId: createClientNote
      tags: [clients]
      summary: Add note to client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientNote"
  /clients/{id}/followups:
    get:
      operationId: listClientFollowups
      tags: [clients]
      summary: List follow-ups for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowUp
      tags: [clients]
      summary: Schedule a follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /followups:
    get:
      operationId: listAllFollowups
      tags: [clients]
      summary: List all upcoming follow-ups
      parameters:
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: dueBefore
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"

  # ── HOTELS DATABASE ────────────────────────────────────────────────
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Get monthly revenue data
      parameters:
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────
...
    Client:
      type: object
      required: [id, name, email, phone, country, leadStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        leadStatus:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        totalBookings:
          type: integer
        createdAt:
          type: string
    ClientDetail:
      type: object
      required: [id, name, email, phone, country, leadStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        leadStatus:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ClientNote"
        followUps:
          type: array
          items:
            $ref: "#/components/schemas/FollowUp"
        quotations:
          type: array
          items:
            $ref: "#/components/schemas/Quotation"
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, email, phone, country]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        country:
          type: string
        city:
          type: string
        leadStatus:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        assignedTo:
          type: integer
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        country:
          type: string
        city:
          type: string
        leadStatus:
          type: string
        assignedTo:
          type: integer
    ClientNote:
      type: object
      required: [id, clientId, content, createdBy, createdByName, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        content:
          type: string
        createdBy:
          type: integer
        createdByName:
          type: string
        createdAt:
          type: string
    ClientNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    FollowUp:
      type: object
      required: [id, clientId, clientName, dueDate, type, status, assignedTo, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        dueDate:
          type: string
        type:
          type: string
          enum: [call, email, whatsapp, meeting]
        status:
          type: string
          enum: [pending, completed, cancelled]
        notes:
          type: ["string", "null"]
        assignedTo:
          type: integer
        assignedToName:
          type: ["string", "null"]
        createdAt:
          type: string
    FollowUpInput:
      type: object
      required: [dueDate, type, assignedTo]
      properties:
        dueDate:
          type: string
        type:
          type: string
          enum: [call, email, whatsapp, meeting]
        notes:
          type: string
        assignedTo:
          type: integer

    # ── HOTELS ─────────────────────────────────────────────────────
...
    Quotation:
      type: object
      required: [id, clientId, clientName, referenceNo, status, totalAmount, currency, validUntil, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        referenceNo:
          type: string
        title:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, accepted, rejected, expired]
        totalAmount:
          type: number
        currency:
          type: string
        validUntil:
          type: string
        createdBy:
          type: integer
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ActivityItem:
      type: object
      required: [id, type, description, entityType, entityId, userId, userName, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
    RevenueDataPoint:
      type: object
      required: [month, revenue, expenses, profit]
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
...